Edit Listing now working
edit listing does not update firebase yet (only changes info on ViewOwnListing but does not change the actual listing in userListing. I think the home fragment need to pull from firebase so it will be updated)

fixed back button on ViewOwnListing so it goes back to MainActivity instead of previous activity.

also for the price attribute of listing I'm just gonna leave it as a string because it's easier to convert to bigdecimal and setting/getting text i guess?